# oj-backend/src/main/java/com/binzc/oj/model/dto/user/UserLoginRequest.java
package com.binzc.oj.model.dto.user;

import java.io.Serializable;
import lombok.Data;

/**
 * 用户登录请求
 *
 * @author binzc
 * @Contact 2019500261@qq.com
 */
@Data
public class UserLoginRequest implements Serializable {

    private static final long serialVersionUID = 3191241716373120793L;

    private String userAccount;

    private String userPassword;
}