# oj-frontend/api-docs.json
{"swagger":"2.0","info":{"description":"提供题目管理、用户管理等功能的 API","version":"1.0","title":"OJ 后端 API"},"host":"localhost:8121","tags":[{"name":"question-controller","description":"Question Controller"},{"name":"question-submit-controller","description":"Question Submit Controller"},{"name":"user-controller","description":"User Controller"}],"paths":{"/api/question/add":{"post":{"tags":["question-controller"],"summary":"addQuestion","operationId":"addQuestionUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"questionAddRequest","description":"questionAddRequest","required":true,"schema":{"originalRef":"QuestionAddRequest","$ref":"#/definitions/QuestionAddRequest"}}],"responses":{"200":{"description":"OK","schema":{"originalRef":"BaseResponse«long»","$ref":"#/definitions/BaseResponse«long»"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/question/detail/{id}":{"get":{"tags":["question-controller"],"summary":"getQuestionById","operationId":"getQuestionByIdUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"originalRef":"BaseResponse«QuestionDetailSafeVo»","$ref":"#/definitions/BaseResponse«QuestionDetailSafeVo»"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/question/list":{"get":{"tags":["question-controller"],"summary":"listQuestion","operationId":"listQuestionUsingGET","produces":["*/*"],"parameters":[{"name":"name","in":"query","description":"name","required":false,"type":"string"},{"name":"page","in":"query","description":"page","required":false,"type":"integer","default":1,"format":"int32"},{"name":"size","in":"query","description":"size","required":false,"type":"integer","default":10,"format":"int32"},{"name":"tags","in":"query","description":"tags","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"originalRef":"BaseResponse«QuestionListSafeVo»","$ref":"#/definitions/BaseResponse«QuestionListSafeVo»"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/question/submit/":{"post":{"tags":["question-submit-controller"],"summary":"doQuestionSubmit","operationId":"doQuestionSubmitUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"questionSubmitAddRequest","description":"questionSubmitAddRequest","required":true,"schema":{"originalRef":"QuestionSubmitAddRequest","$ref":"#/definitions/QuestionSubmitAddRequest"}}],"responses":{"200":{"description":"OK","schema":{"originalRef":"BaseResponse«QuestionSubmitVo»","$ref":"#/definitions/BaseResponse«QuestionSubmitVo»"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/question/submit/getCode":{"get":{"tags":["question-submit-controller"],"summary":"getCode","operationId":"getCodeUsingGET","produces":["*/*"],"parameters":[{"name":"questionId","in":"query","description":"questionId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"originalRef":"BaseResponse«CodeVo»","$ref":"#/definitions/BaseResponse«CodeVo»"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/question/submit/record/{id}":{"get":{"tags":["question-submit-controller"],"summary":"getRecord","operationId":"getRecordUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"originalRef":"BaseResponse«SubmitRecordDetail»","$ref":"#/definitions/BaseResponse«SubmitRecordDetail»"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/question/submit/submitRecord":{"post":{"tags":["question-submit-controller"],"summary":"submitRecords","operationId":"submitRecordsUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"queryParmRequest","description":"queryParmRequest","required":true,"schema":{"originalRef":"QueryParmRequest","$ref":"#/definitions/QueryParmRequest"}}],"responses":{"200":{"description":"OK","schema":{"originalRef":"BaseResponse«SubmitRecodWithPageVo»","$ref":"#/definitions/BaseResponse«SubmitRecodWithPageVo»"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/user/confirmChangeAvatar":{"post":{"tags":["user-controller"],"summary":"confirmChangeAvatar","operationId":"confirmChangeAvatarUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"userConfirmChangeAvatar","description":"userConfirmChangeAvatar","required":true,"schema":{"originalRef":"UserConfirmChangeAvatar","$ref":"#/definitions/UserConfirmChangeAvatar"}}],"responses":{"200":{"description":"OK","schema":{"originalRef":"BaseResponse","$ref":"#/definitions/BaseResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/user/get/login":{"get":{"tags":["user-controller"],"summary":"getLoginUser","operationId":"getLoginUserUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"originalRef":"BaseResponse«LoginUserVO»","$ref":"#/definitions/BaseResponse«LoginUserVO»"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/user/hello":{"get":{"tags":["user-controller"],"summary":"hello","operationId":"helloUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/user/login":{"post":{"tags":["user-controller"],"summary":"userLogin","operationId":"userLoginUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"userLoginRequest","description":"userLoginRequest","required":true,"schema":{"originalRef":"UserLoginRequest","$ref":"#/definitions/UserLoginRequest"}}],"responses":{"200":{"description":"OK","schema":{"originalRef":"BaseResponse«LoginUserVO»","$ref":"#/definitions/BaseResponse«LoginUserVO»"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/user/register":{"post":{"tags":["user-controller"],"summary":"userRegister","operationId":"userRegisterUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"userRegisterRequest","description":"userRegisterRequest","required":true,"schema":{"originalRef":"UserRegisterRequest","$ref":"#/definitions/UserRegisterRequest"}}],"responses":{"200":{"description":"OK","schema":{"originalRef":"BaseResponse«long»","$ref":"#/definitions/BaseResponse«long»"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/user/uploadAvatar":{"post":{"tags":["user-controller"],"summary":"uploadAvatar","operationId":"uploadAvatarUsingPOST","consumes":["multipart/form-data"],"produces":["*/*"],"parameters":[{"name":"file","in":"formData","required":false,"type":"file"}],"responses":{"200":{"description":"OK","schema":{"originalRef":"BaseResponse«string»","$ref":"#/definitions/BaseResponse«string»"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}}},"definitions":{"BaseResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"object"},"message":{"type":"string"}},"title":"BaseResponse"},"BaseResponse«CodeVo»":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"originalRef":"CodeVo","$ref":"#/definitions/CodeVo"},"message":{"type":"string"}},"title":"BaseResponse«CodeVo»"},"BaseResponse«LoginUserVO»":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"originalRef":"LoginUserVO","$ref":"#/definitions/LoginUserVO"},"message":{"type":"string"}},"title":"BaseResponse«LoginUserVO»"},"BaseResponse«QuestionDetailSafeVo»":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"originalRef":"QuestionDetailSafeVo","$ref":"#/definitions/QuestionDetailSafeVo"},"message":{"type":"string"}},"title":"BaseResponse«QuestionDetailSafeVo»"},"BaseResponse«QuestionListSafeVo»":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"originalRef":"QuestionListSafeVo","$ref":"#/definitions/QuestionListSafeVo"},"message":{"type":"string"}},"title":"BaseResponse«QuestionListSafeVo»"},"BaseResponse«QuestionSubmitVo»":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"originalRef":"QuestionSubmitVo","$ref":"#/definitions/QuestionSubmitVo"},"message":{"type":"string"}},"title":"BaseResponse«QuestionSubmitVo»"},"BaseResponse«SubmitRecodWithPageVo»":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"originalRef":"SubmitRecodWithPageVo","$ref":"#/definitions/SubmitRecodWithPageVo"},"message":{"type":"string"}},"title":"BaseResponse«SubmitRecodWithPageVo»"},"BaseResponse«SubmitRecordDetail»":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"originalRef":"SubmitRecordDetail","$ref":"#/definitions/SubmitRecordDetail"},"message":{"type":"string"}},"title":"BaseResponse«SubmitRecordDetail»"},"BaseResponse«long»":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"integer","format":"int64"},"message":{"type":"string"}},"title":"BaseResponse«long»"},"BaseResponse«string»":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"string"},"message":{"type":"string"}},"title":"BaseResponse«string»"},"CodeVo":{"type":"object","properties":{"code":{"type":"string"},"language":{"type":"string"}},"title":"CodeVo"},"JudgeCase":{"type":"object","properties":{"input":{"type":"string"},"output":{"type":"string"}},"title":"JudgeCase"},"JudgeConfig":{"type":"object","properties":{"memoryLimit":{"type":"integer","format":"int64"},"stackLimit":{"type":"integer","format":"int64"},"timeLimit":{"type":"integer","format":"int64"}},"title":"JudgeConfig"},"JudgeInfo":{"type":"object","properties":{"memory":{"type":"integer","format":"int64"},"message":{"type":"string"},"time":{"type":"integer","format":"int64"}},"title":"JudgeInfo"},"JudgeMessage":{"type":"object","properties":{"memory":{"type":"integer","format":"int64"},"status":{"type":"string"},"statusSingle":{"type":"string"},"time":{"type":"integer","format":"int64"}},"title":"JudgeMessage"},"LoginUserVO":{"type":"object","properties":{"createTime":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int64"},"updateTime":{"type":"string","format":"date-time"},"userAvatar":{"type":"string"},"userName":{"type":"string"},"userProfile":{"type":"string"},"userRole":{"type":"string"}},"title":"LoginUserVO"},"QueryParmRequest":{"type":"object","properties":{"language":{"type":"string"},"pageNo":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"questionTitle":{"type":"string"},"submitStatus":{"type":"integer","format":"int32"},"userName":{"type":"string"}},"title":"QueryParmRequest"},"QuestionAddRequest":{"type":"object","properties":{"answer":{"type":"string"},"content":{"type":"string"},"judgeCase":{"type":"array","items":{"originalRef":"JudgeCase","$ref":"#/definitions/JudgeCase"}},"judgeConfig":{"originalRef":"JudgeConfig","$ref":"#/definitions/JudgeConfig"},"tags":{"type":"array","items":{"type":"string"}},"title":{"type":"string"}},"title":"QuestionAddRequest"},"QuestionDetailSafeVo":{"type":"object","properties":{"acceptedNum":{"type":"integer","format":"int32"},"answer":{"type":"string"},"content":{"type":"string"},"id":{"type":"integer","format":"int64"},"judgeConfig":{"originalRef":"JudgeConfig","$ref":"#/definitions/JudgeConfig"},"submitNum":{"type":"integer","format":"int32"},"tags":{"type":"array","items":{"type":"string"}},"title":{"type":"string"}},"title":"QuestionDetailSafeVo"},"QuestionListSafeVo":{"type":"object","properties":{"currentPage":{"type":"integer","format":"int64"},"pageSize":{"type":"integer","format":"int64"},"questions":{"type":"array","items":{"originalRef":"QuestionSafeVo","$ref":"#/definitions/QuestionSafeVo"}},"totalPages":{"type":"integer","format":"int64"},"totalRecords":{"type":"integer","format":"int64"}},"title":"QuestionListSafeVo"},"QuestionSafeVo":{"type":"object","properties":{"acceptedNum":{"type":"integer","format":"int32"},"id":{"type":"integer","format":"int64"},"submitNum":{"type":"integer","format":"int32"},"tags":{"type":"string"},"title":{"type":"string"}},"title":"QuestionSafeVo"},"QuestionSubmitAddRequest":{"type":"object","properties":{"code":{"type":"string"},"language":{"type":"string"},"questionId":{"type":"integer","format":"int64"}},"title":"QuestionSubmitAddRequest"},"QuestionSubmitVo":{"type":"object","properties":{"code":{"type":"string"},"createTime":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int64"},"judgeInfo":{"originalRef":"JudgeInfo","$ref":"#/definitions/JudgeInfo"},"language":{"type":"string"},"loginUserVO":{"originalRef":"LoginUserVO","$ref":"#/definitions/LoginUserVO"},"questionDetailSafeVo":{"originalRef":"QuestionDetailSafeVo","$ref":"#/definitions/QuestionDetailSafeVo"},"questionId":{"type":"integer","format":"int64"},"status":{"type":"integer","format":"int32"},"updateTime":{"type":"string","format":"date-time"},"userId":{"type":"integer","format":"int64"}},"title":"QuestionSubmitVo"},"SubmitRecodWithPageVo":{"type":"object","properties":{"pageNo":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"submitRecordSimples":{"type":"array","items":{"originalRef":"SubmitRecordSimple","$ref":"#/definitions/SubmitRecordSimple"}},"total":{"type":"integer","format":"int64"}},"title":"SubmitRecodWithPageVo"},"SubmitRecordDetail":{"type":"object","properties":{"code":{"type":"string"},"createTime":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int64"},"judgeInfo":{"originalRef":"JudgeInfo","$ref":"#/definitions/JudgeInfo"},"judgeMessages":{"type":"array","items":{"originalRef":"JudgeMessage","$ref":"#/definitions/JudgeMessage"}},"language":{"type":"string"},"message":{"type":"string"},"questionId":{"type":"integer","format":"int64"},"questionTitle":{"type":"string"},"status":{"type":"integer","format":"int32"},"userAvatar":{"type":"string"},"userId":{"type":"integer","format":"int64"},"userName":{"type":"string"}},"title":"SubmitRecordDetail"},"SubmitRecordSimple":{"type":"object","properties":{"codeLength":{"type":"string"},"judgeInfo":{"type":"string"},"language":{"type":"string"},"memoryUse":{"type":"integer","format":"int64"},"message":{"type":"string"},"questionId":{"type":"integer","format":"int64"},"questionTitle":{"type":"string"},"status":{"type":"integer","format":"int32"},"submitId":{"type":"integer","format":"int64"},"submitTime":{"type":"string","format":"date-time"},"timeUse":{"type":"integer","format":"int64"},"userAvatar":{"type":"string"},"userId":{"type":"string"},"userName":{"type":"string"}},"title":"SubmitRecordSimple"},"UserConfirmChangeAvatar":{"type":"object","properties":{"url":{"type":"string"},"yesOrNo":{"type":"boolean"}},"title":"UserConfirmChangeAvatar"},"UserLoginRequest":{"type":"object","properties":{"userAccount":{"type":"string"},"userPassword":{"type":"string"}},"title":"UserLoginRequest"},"UserRegisterRequest":{"type":"object","properties":{"checkPassword":{"type":"string"},"userAccount":{"type":"string"},"userPassword":{"type":"string"},"userRole":{"type":"string"}},"title":"UserRegisterRequest"}}}